+++ src/OrchardCore.Modules/OrchardCore.ContentPreview/Migrations.cs
@@ -1,0 +1,25 @@
+using OrchardCore.ContentManagement.Metadata;
+using OrchardCore.ContentManagement.Metadata.Settings;
+using OrchardCore.Data.Migration;
+
+namespace OrchardCore.ContentPreview
+{
+    public class Migrations : DataMigration
+    {
+        IContentDefinitionManager _contentDefinitionManager;
+
+        public Migrations(IContentDefinitionManager contentDefinitionManager)
+        {
+            _contentDefinitionManager = contentDefinitionManager;
+        }
+
+        public int Create()
+        {
+            _contentDefinitionManager.AlterPartDefinition("PreviewPart", builder => builder
+                .Attachable()
+                .WithDescription("Provides a way to define the url that is used to render your content item."));
+
+            return 1;
+        }
+    }
+}
