+++ src/OrchardCore.Modules/OrchardCore.ContentPreview/OrchardCore.ContentPreview.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\..\OrchardCore\OrchardCore.ContentManagement.Display\OrchardCore.ContentManagement.Display.csproj" />
     <ProjectReference Include="..\..\OrchardCore\OrchardCore.ResourceManagement\OrchardCore.ResourceManagement.csproj" />
 
+    <ProjectReference Include="..\OrchardCore.Contents\OrchardCore.Contents.csproj" />
     <ProjectReference Include="..\OrchardCore.Settings\OrchardCore.Settings.csproj" />
   </ItemGroup>
