+++ src/OrchardCore.Modules/OrchardCore.ContentPreview/Pages/_ViewImports.cshtml
@@ -1,0 +1,5 @@
+@inherits OrchardCore.DisplayManagement.RazorPages.Page
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, OrchardCore.DisplayManagement
+@addTagHelper *, OrchardCore.ResourceManagement
+@addTagHelper *, OrchardCore.Contents
